requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
pyodbc
openpyxl
supabase
azure-storage-blob
pandas
azure-functions-worker